### 2024-11-19 #7502 @Kevin-Mattheus-Moerman
@ljwoods2 - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
