+++ joss_paper/paper.bib
@@ -57,7 +57,7 @@
     pages = {1546-1553},
     year = {2014},
     issn = {0010-4655},
-    doi = {https://doi.org/10.1016/j.cpc.2014.01.018},
+    doi = {10.1016/j.cpc.2014.01.018},
     url = {https://www.sciencedirect.com/science/article/pii/S0010465514000447},
     author = {Pierre {de Buyl} and Peter H. Colberg and Felix Höfling},
     keywords = {Molecular simulation, HDF5},
@@ -89,7 +89,8 @@
   pages      = "76--84",
   year       =  2023,
   conference = "Python in Science Conference",
-  location   = "Austin, Texas"
+  location   = "Austin, Texas",
+  doi = {10.25080/gerudo-f2bc6f59-00a}
 }
 
 
@@ -112,7 +113,7 @@
     number = {10},
     pages = {2319-2327},
     keywords = {molecular dynamics simulations, analysis, proteins, object-oriented design, software, membrane systems, Python programming language},
-    doi = {https://doi.org/10.1002/jcc.21787},
+    doi = {10.1002/jcc.21787},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
     abstract = {Abstract MDAnalysis is an object-oriented library for structural and temporal analysis of molecular dynamics (MD) simulation trajectories and individual protein structures. It is written in the Python language with some performance-critical code in C. It uses the powerful NumPy package to expose trajectory data as fast and efficient NumPy arrays. It has been tested on systems of millions of particles. Many common file formats of simulation packages including CHARMM, Gromacs, Amber, and NAMD and the Protein Data Bank format can be read and written. Atoms can be selected with a syntax similar to CHARMM's powerful selection commands. MDAnalysis enables both novice and experienced programmers to rapidly write their own analytical tools and access data stored in trajectories in an easily accessible manner that facilitates interactive explorative analysis. MDAnalysis has been tested on and works for most Unix-based platforms such as Linux and Mac OS X. It is freely available under the GNU General Public License from http://mdanalysis.googlecode.com. © 2011 Wiley Periodicals, Inc. J Comput Chem 2011},
@@ -127,6 +128,7 @@
       archivePrefix={arXiv},
       primaryClass={q-bio.BM},
       url={https://arxiv.org/abs/2407.16584}, 
+      doi={10.48550/arXiv.2407.16584}
 }
 
 @article{MDsrv:2022,
@@ -278,7 +280,7 @@ title = "{Molecular Dynamics trajectories of membrane protein YiiP}",
 year = "2019",
 month = "5",
 url = "https://figshare.com/articles/dataset/Molecular_Dynamics_trajectories_of_membrane_protein_YiiP/8202149",
-doi = "10.6084/m9.figshare.8202149.v1"
+doi = {10.6084/m9.figshare.8202149.v1}
 }
 
 @misc{Zarr:2024,
@@ -315,4 +317,6 @@ doi = "10.6084/m9.figshare.8202149.v1"
 	pages = {1561-1563},
 	title = {Fast determination of the optimal rotational matrix for macromolecular superpositions},
 	volume = 31,
-	year = 2010}
+	year = 2010,
+  doi = {10.1002/jcc.21439},  
+}
