### 2024-12-07 Becksteinlab/zarrtraj#73 @ljwoods2
JOSS Submission
Standing issue for the JOSS PR:

https://github.com/openjournals/joss-reviews/issues/7502
