+++ README.md
@@ -5,20 +5,20 @@ zarrtraj
 | **Latest release** | [![Last release tag][badge_release]][url_latest_release] ![GitHub commits since latest release (by date) for a branch][badge_commits_since]  [![Documentation Status][badge_docs]][url_docs]|
 | :----------------- | :------- |
 | **Status**         | [![GH Actions Status][badge_actions]][url_actions] [![codecov][badge_codecov]][url_codecov] |
-| **Community**      | [![License: GPL v3][badge_license]][url_license]  [![Powered by MDAnalysis][badge_mda]][url_mda]|
+| **Community**      | [![License: MIT][badge_license]][url_license]  [![Powered by MDAnalysis][badge_mda]][url_mda]|
 
 [badge_actions]: https://github.com/Becksteinlab/zarrtraj/actions/workflows/gh-ci.yaml/badge.svg
 [badge_codecov]: https://codecov.io/gh/Becksteinlab/zarrtraj/branch/main/graph/badge.svg
 [badge_commits_since]: https://img.shields.io/github/commits-since/Becksteinlab/zarrtraj/latest
 [badge_docs]: https://readthedocs.org/projects/zarrtraj/badge/?version=latest
-[badge_license]: https://img.shields.io/badge/License-GPLv3-blue.svg
+[badge_license]: https://img.shields.io/badge/License-MIT-blue.svg
[…]
