+++ CITATION.cff
@@ -62,4 +62,4 @@ keywords:
   - file-format
   - mdanalysis
   - zarr
-license: GPL-3.0-or-later
+license: MIT
