### 2025-02-05 #7502 @RMeli
Hi @Kevin-Mattheus-Moerman, thank you for the invite to edit this submission. Unfortunately I have a conflict of interest for this paper. I'm a core developer of MDAnalysis, and so are 4 of the authors of the paper.
