### 2025-02-15 Becksteinlab/zarrtraj#73 @orbeckst
Just fyi, they are still looking for a handling editor for the paper.
