### 2025-03-24 #7943 @editorialbot
[REVIEW]: Zarrtraj: A Python package for streaming molecular dynamics trajectories from cloud services
**Submitting author:** @ljwoods2 (<a href="http://orcid.org/0009-0003-0713-4167">Lawson Woods</a>)
**Repository:** https://github.com/Becksteinlab/zarrtraj
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
