### 2025-07-14 1868f0a
pin zarr<3.0
