### 2025-07-14 594589a
pin zarr<3.0
