+++ devtools/conda-envs/test_env.yaml
@@ -9,7 +9,7 @@ dependencies:
 
   ### Core requirements ###
   - MDAnalysis>=2.1.0
-  - zarr>=2.11.0
+  - zarr>=2.11.0,<3.0
   - kerchunk>=0.2.6
   - h5py>=3.11.0
 
