+++ .github/workflows/gh-ci.yaml
@@ -50,7 +50,7 @@ jobs:
           - macOS-latest
           - ubuntu-latest
           - windows-latest
-        python-version: ["3.10", "3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
         mdanalysis-version:
           - latest
           - develop
