### 2025-07-18 #7943 @srmnitc
@ljwoods2 could you please take a look at these points? Thanks!

> I would like to raise two points:
>
>     * Do I see it correctly that there is no ready to run example on the GitHub page? Do you think there's a possibility to test it with `gcs://zarrtraj-test-data` cloud bucket as indicated in the pdf file?
>
>     * I don't quite understand what `import zarrtraj` is actually doing, because `zarrtraj` does not appear in the code. Could you explain this part a bit more? Does the order of imports matter?
