### 2025-07-20 c346250
doc improvements
