+++ docs/source/walkthrough.rst
@@ -1,18 +1,31 @@
+.. _walkthrough:
+
 Walkthrough
 ===========
 
-This walkthrough will guide you through the process of reading and writing H5MD-formatted trajectories from cloud storage using 
+This walkthrough will guide you through the process of writing and then reading H5MD-formatted trajectories from cloud storage using 
 AWS S3 as an example. To learn more about reading and writing trajectories from different cloud storage providers, 
-including Google Cloud and Azure, see the API documentation.
+including Google Cloud and Azure, see the :ref:`API documentation <_api>`.
 
-Reading H5MD trajectories from cloud storage
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+We will use the `YiiP 9ns trajectory from MDAnalysisData <https://www.mdanalysis.org/MDAnalysisData/yiip_equilibrium.html>`
+as an example trajectory for the walkthrough.
 
-Uploading your H5MD file
-########################
+.. note:: 
+    In examples that read or write the :ref:`ZarrMD format <_zarrmd>`, `zarrtraj` is imported even though 
+    its namespace is not explicitly called. This is because MDAnalysis uses `meta classes for registering reader and writer formats <https://github.com/MDAnalysis/mdanalysis/blob/d412c9a9a56312c1bd4e33e6dd3afc4cec7783ca/package/MDAnalysis/coordinates/base.py>`,
+    so calling `import zarrtraj` allows MDAnalysis to use the classes provided by this package.
 
-First, upload your H5MD trajectories to an AWS S3 bucket. This requires that an S3 Bucket is setup and configured for 
-write access using the credentials stored in "sample_profile". If you've never configured an S3 Bucket before, see
+Writing trajectories to cloud storage
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+In this section, we show how to upload both H5MD files and ZarrMD files to an S3 bucket.
+
+Setup: Creating an AWS S3 bucket
+################################
+
+First, create an AWS S3 bucket. This requires that an S3 Bucket is setup and configured for 
+write access using the credentials stored in "sample_profile" (for the examples shown here, but a profile can be named arbitrarily). 
+If you've never configured an S3 Bucket before, see
 `this guide <https://docs.aws.amazon.com/AmazonS3/latest/userguide/creating-bucket.html>`_. You can setup a profile to easily manage AWS
 credentials using `this VSCode extension <https://marketplace.visualstudio.com/items?itemName=AmazonWebServices.aws-toolkit-vscode>`_.
 Here is a sample profile (stored in ~/.aws/credentials) where 
@@ -22,15 +35,62 @@ Here is a sample profile (stored in ~/.aws/credentials) where
     [sample_profile]
     aws_access_key_id = <key>
 
+
+Writing trajectories directly from MDAnalysis into a ZarrMD file in an S3 Bucket (preferred method)
+####################################################################################################
+
+Using the created credentials with read/write access, you can write a trajectory
+into your bucket.
+
+You can change the stored precision of floating point values in the trajectory file with the optional
+``precision`` kwarg and pass in any ```numcodecs.Codec``` compressor with the optional
+``compressor`` kwarg. See `numcodecs <https://numcodecs.readthedocs.io/en/stable/>`_
+for more on the available compressors.
+
+Chunking is automatically determined for all datasets to be optimized for
+cloud storage reading speed and is not configurable by the user::
+
[…]
