### 2025-07-20 e46e923
typo in API doc
