### 2025-07-20 4ba5cf1
clarify traj accessibility
