### 2025-07-21 65965d6
doc additions for beginner users
