### 2025-07-21 2e8ddbc
Update yiip_example.rst
