### 2025-07-21 #7943 @ljwoods2
@samwaseda Thank you for the feedback, your point is well-taken. Worked with the other authors to improve our documentation and make some directly runnable examples:

- A jupyter notebook where we perform [RMSD on the openly available GCS YiiP trajectory](https://github.com/Becksteinlab/zarrtraj/blob/main/examples/rmsd_yiip.ipynb), which is directly runnable and produces a plot
- Improved explanation of example code (including need for importing zarrtraj) and addition of some code that actually [interacts with the trajectory](https://zarrtraj.readthedocs.io/en/latest/yiip_example.html), runnable if copy-and-pasted
- Rewrote the [docs on storing a trajectory on the cloud yourself](https://zarrtraj.readthedocs.io/en/latest/walkthrough.html) with a more clear ordering and changed everything to use the same trajectory

It's difficult to directly make the code that we used to produce the figures in the paper runnable by anyone, just because storing and making the longer trajectory freely streamable on S3 costs more than would be free in S3 free tier (Google's free tier is a bit more generous, which is how we were able to at least host a single trajectory for free on it)- but if that's a major concern, please let me know and we can potentially find at least a temporary workaround
