### 2025-09-08 #7943 @mattwthompson
Hi @ljwoods2 I am impressed with both the software and paper and, with the help of other recent improvements, I'm able to easily find and run through the example hosted on GCS which gives me confidence in the general performance claims of the work. I'm having a hard time, however, finding a way to reproduce parts of Fig. 1 in which comparison to slower methods (on the full trajectory) are used.

I don't need to fully reproduce existing benchmarks but I would like to compare to one or two slower methods. Is the best way for me to do this by doing the following?
1. Use [this](https://www.mdanalysis.org/MDAnalysisData/yiip_equilibrium.html) link to grab the 4 GB PDB file
2. Use [these scripts](https://github.com/Becksteinlab/zarrtraj/tree/main/zarrtraj/data) to convert that file between formats
3. Re-run parts of [this notebok](https://github.com/Becksteinlab/zarrtraj/blob/main/joss_paper/figure_2.ipynb) to evaluate the (relative) performance claims

I anticipate the claims are true but would like to have that "wow" moment running the comparisons myself to be sure 🙂
