### 2025-09-09 #7943 @samwaseda
I'm VERY sorry that it's taking so long. Now I should have more time in the next couple of weeks, so I promise to have done the full checks soon.

I wanted to run [this notebook](https://github.com/Becksteinlab/zarrtraj/blob/main/examples/rmsd_yiip.ipynb), but it looks like I don't have access rights:

```
OSError: Forbidden: https://storage.googleapis.com/download/storage/v1/b/zarrtraj-test-data/o/YiiP_system.pdb?alt=media
The billing account for the owning project is disabled in state delinquent
```

Do you think you can make `gcs://zarrtraj-test-data/YiiP_system.pdb` public?
