### 2025-09-16 00e0ec6
typo in fetch code
