### 2025-09-16 #7943 @ljwoods2
@mattwthompson Thank you, and you're right that those are the overall steps the (with the note that the 4 GB file is an XTC trajectory file, and the topology PDB file is much smaller)

Let me write a more explicit set of steps to run the benchmarks or perform something similar to them, I appreciate you pointing this out
