### 2025-10-08 #7943 @ljwoods2
@srmnitc Great timing, just finished:

Added descriptions of how to run stride time and RMSD benchmarks directly in the figure notebooks

- https://github.com/Becksteinlab/zarrtraj/blob/main/joss_paper/figure_1.ipynb
- https://github.com/Becksteinlab/zarrtraj/blob/main/joss_paper/figure_2.ipynb
