### 2025-11-21 #7943 @samwaseda
Oh sorry I didn't realize that the notebook had been updated. I checked it now and from my side it's ready to go!
