### 2026-01-28 #7943 @mattwthompson
Actively working on it now, just making some novice mistakes setting up my S3 bucket. Hope to run the notebook fully through soon!
