### 2026-02-24 #7943 @mattwthompson
This one's still on me - but I'm happy to report I've been able to go through the performance benchmarks and reproduce, at a high level, the performance benchmarks in the paper. I'll go back through the checklist shortly

What I get (maybe on quicker hardware? not sure):

What the authors reported:

In any case this demonstrates the key value add of the software, that it outperforms H5MD on disk and S3
