### 2026-02-24 #7943 @mattwthompson
So that ticks off the reproducibility and performance boxes in [my checklist](https://github.com/openjournals/joss-reviews/issues/7943#issuecomment-3067777131).

I think that all adds up to a thumbs-up from me
