### 2026-02-28 #7943 @ljwoods2
@srmnitc

v0.3.3 on Zenodo (10.5281/zenodo.18819459)
https://zenodo.org/records/18819459
