### 2026-03-07 #7943 @srmnitc
@ljwoods2 There is a slight mismatch on the names in the paper and the repo: Richard Gowers and Richard J Gowers, could you please make it consistent?
