### 2026-03-12 #7943 @ljwoods2
@srmnitc Updated it on the Zenodo! Can you please advise on https://github.com/Becksteinlab/zarrtraj/issues/91 what can be done about modifying the rendering of JOSS style citations- it seems this can't be changed?
