### 2026-03-13 #7943 @srmnitc
> > [@ljwoods2](https://github.com/ljwoods2) There is a slight mismatch on the names in the paper and the repo: Richard Gowers and Richard J Gowers, could you please make it consistent?
>
> [@srmnitc](https://github.com/srmnitc) , could you please be clear about where exactly author names do not match? Maybe I am just misunderstanding your request?
>
> "Richard J Gowers" is how [@richardjgowers](https://github.com/richardjgowers) lists his name in all his academic papers, see https://scholar.google.com/citations?user=ujKb5B8AAAAJ ; the zarrtraj [CITATION.cff](https://github.com/Becksteinlab/zarrtraj/blob/e85d3c8b6d5db36937aa1b4c6771f685655895b8/CITATION.cff#L38-L43) lists him as "Richard J Gowers". I am a bit confused as to what we should be changing in the current paper or repo because every instance where we can write out the author name is consistent between repo and paper.
>
> Short of [@richardjgowers](https://github.com/richardjgowers) stating here that he wants to change how he appears as an author on the paper I don't see how we can change anything.
>
> GitHub handles do not have to be associated with real names so I don't see that the author list needs to agree with those.
>
> Thanks!

@orbeckst the mismatch was between the Zenodo archive and the paper. The Zenodo archive was missing 'J'. But @ljwoods2 has fixed it already, thanks!
