### 2026-03-27 #7943 @srmnitc
> > [@srmnitc](https://github.com/srmnitc) Updated it on the Zenodo! Can you please advise on [Becksteinlab/zarrtraj#91](https://github.com/Becksteinlab/zarrtraj/issues/91) what can be done about modifying the rendering of JOSS style citations- it seems this can't be changed?
>
> [@rkurchin](https://github.com/rkurchin) we have a small issue here that the references are collapsed with dots when there is a large number of authors. Is this fine, or should we make any change?

@Kevin-Mattheus-Moerman could you please help with this - we are ready to finish up here!
