### 2026-03-28 #7943 @Kevin-Mattheus-Moerman
@openjournals/dev can you have a look at the use of `...` in JOSS paper references when there are many authors. I take it this is intentional? Can we control this behaviour or is it authomatic. I do not think it is a problem actually, just checking.
