### 2026-03-28 #7943 @Kevin-Mattheus-Moerman
@srmnitc @ljwoods2 on the Zenodo archive link and the spelling of names like Richard Gowers vs Richard J Gowers. I think it is good to try to adhere to an authors preferred spelling, especially of course on the JOSS paper. However, in principle, for JOSS, I feel it is not so important that the two match exactly with the archive. It looks like it is solved now already, and we can proceed.

@srmnitc I've pinged our dev team about the `...` thing. However, I think unless it is unwanted behaviour here, it is probably set as a deliberate setting. In other words I think you can proceed here and we should get a response from the dev team shortly.
