### 2026-03-28 #7943 @arfon
> @srmnitc I've pinged our dev team about the ... thing. However, I think unless it is unwanted behaviour here, it is probably set as a deliberate setting. In other words I think you can proceed here and we should get a response from the dev team shortly.

This is not something in our control (library behavior of a dependency) so I suggest we proceed.
