### 2026-04-01 #7943 @ljwoods2
@srmnitc I created a new github and zenodo release for the typos you found:

https://zenodo.org/records/19363058
