### 2026-04-16 #7943 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
